[#26] Refactor file layout
