+++ redux/src/pages/Home/index.test.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
-import { store } from './app/store';
-import App from './App';
+import { store } from '../../app/store';
+import Home from '.';
 
 test('renders learn react link', () => {
   const { getByText } = render(
     <Provider store={store}>
-      <App />
+      <Home />
     </Provider>
   );
