+++ redux/src/pages/Home/style.css
@@ -1,19 +1,19 @@
-.App {
+.Home {
   text-align: center;
 }
 
-.App-logo {
+.Home-logo {
   height: 40vmin;
   pointer-events: none;
 }
 
 @media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-float infinite 3s ease-in-out;
+  .Home-logo {
+    animation: Home-logo-float infinite 3s ease-in-out;
   }
 }
 
-.App-header {
+.Home-header {
   min-height: 100vh;
   display: flex;
   flex-direction: column;
@@ -22,11 +22,11 @@
   font-size: calc(10px + 2vmin);
 }
 
-.App-link {
+.Home-link {
   color: rgb(112, 76, 182);
 }
 
-@keyframes App-logo-float {
+@keyframes Home-logo-float {
   0% {
     transform: translateY(0);
   }
