+++ redux/src/slices/counter/index.spec.ts
@@ -3,7 +3,7 @@ import counterReducer, {
   increment,
   decrement,
   incrementByAmount,
-} from './counterSlice';
+} from '.';
 
 describe('counter reducer', () => {
   const initialState: CounterState = {
